+++ protocols/kad/CHANGELOG.md
@@ -1,3 +1,20 @@
+# 0.29.0 [unreleased]
+
+- Update `libp2p-swarm`.
+
+# 0.28.1 [2021-02-15]
+
+- Update dependencies.
+
+# 0.28.0 [2021-01-12]
+
+- Update dependencies.
+
+# 0.27.1 [2021-01-11]
+
+- Add From impls for `kbucket::Key`.
+  [PR 1909](https://github.com/libp2p/rust-libp2p/pull/1909).
+
 # 0.27.0 [2020-12-17]
 
 - Update `libp2p-core` and `libp2p-swarm`.